cleaned up <head> markup insertion logic and added to new options to $.mobile hash:
- metaViewportContent: "width=device-width, minimum-scale=1, maximum-scale=1",
- headExtras: undefined

If metaViewportContent is undefined, no viewport meta tag will be added to the dom.

headExtras is undefined by default, but allows for additional markup to be injected into the head element, such as meta tags used for enabling iOS local app features

Also, the <base> tag is defined as $base (undefined if support test is false), so setting its value no longer uses an ID selector to find it.