changed hideBrowserChrome function to "silentScroll". Scrolls to a particular location while disabling scroll event listening.
Function accepts a Y positioning argument, or defaults to 0 if undefined.

Now page transitions cache their previous scroll position when leaving, and that position will be remembered when revisting that page.

This change also includes a fix for webOS to use this new feature and set the previous scroll to the location of the clicked selectmenu (so after using the select, it'll scroll back to its location, even though webOS doesn't support scrolltop)

Fixes #355