Fixes issue #1036 partially offscreen rendering of select menu
https://github.com/jquery/jquery-mobile/issues/issue/1036

The css styles the select menu with a width of 80% with a max-width of
350px. This fix checks if width of select menu is less than max-width.
If that is the case the select menu is centered on screen. If not it
checks if the rigth or left side of the select menu are outside the
viewport and if so moves the select menu inside the viewport with a 30px
tolerance.