Fix for issue 2529 - Transition to the same page
- Added a new allowSamePageTransition option to the changePage() method default settings.

By default, we prevent changePage() requests when the fromPage and toPage are the same element, but folks that generate content manually/dynamically and reuse pages want to be able to transition to the same page. To allow
this, they will need to change the default value of allowSamePageTransition to true, *OR*, pass it in as an option when they manually call changePage().

It should be noted that our default transition animations assume that the formPage and toPage are different elements, so they may behave unexpectedly. It is up to the developer that turns on the allowSamePageTransitiona option
to either turn off transition animations, or make sure that an appropriate animation transition is used.

	// To toggle the default behavior for all changePage() calls,
	// set the default value of allowSamePageTransition to whatever
	// you want it to be. The default is false.

	$.mobile.changePage.defaults.allowSamePageTransition = true;

	// To specify the behavior when manually calling changePage(),
	// pass it as an option. If not specified, the default value
	// specified by $.mobile.changepage.defaults.allowSamePageTransition
	// is used.

	$.mobile.changePage( "#reused-page", { allowSamePageTransition: true } );