Re-vamped the navbar docs to fix incorrect info about icon positioning and theming. Fixes #1858, Fixes #1797, Fixes #2667, Fixes #1180, Fixes #2373
The docs incorrectly stated that the icon position could be set per
link, but this is intended as a global setting for the navbar parent.
This is working as designed and helps with visual consistency. If icon
positioning needs to be be adjusted per link, you will need to manually
apply the right classes through JS and may need to tweak CSS because
top/bottom and left/right positions have different heights.